+++ packages/mcs-lite-mobile-web/src/containers/DeviceDetail/DeviceDetail.js
@@ -16,15 +16,22 @@ import datetimeFormat from '../../utils/datetimeFormat';
 
 class DeviceDetail extends React.Component {
   static propTypes = {
-    device: PropTypes.object,
+    // React-router Params
     deviceId: PropTypes.string.isRequired,
+
+    // Redux State
+    device: PropTypes.object.isRequired,
     isLoading: PropTypes.bool.isRequired,
-    getMessages: PropTypes.func.isRequired,
+
+    // Redux Action
     fetchDeviceDetail: PropTypes.func.isRequired,
-    sendMessage: PropTypes.func,
-  }
-  static defaultProps = {
-    device: {},
+    setDatapoint: PropTypes.func.isRequired,
+
+    // React-intl I18n
+    getMessages: PropTypes.func.isRequired,
+
+    // WebSocket
+    sendMessage: PropTypes.func.isRequired,
   }
   state = { isMenuShow: false, target: undefined };
   componentWillMount = () => this.fetch();
@@ -33,21 +40,24 @@ class DeviceDetail extends React.Component {
   getTarget = node => this.setState({ target: node });
   fetch = () => this.props.fetchDeviceDetail(this.props.deviceId);
   eventHandler = (e) => {
+    const { id, values, type } = e;
+    const { deviceId, sendMessage, setDatapoint } = this.props;
     // TODO: refactor these codes.
-    const datapoint = { datachannelId: e.id, values: e.values };
-    switch (e.type) {
-      case 'submit':
+    const datapoint = { datachannelId: id, values };
+    switch (type) {
+      case 'SUBMIT':
         // Remind: MUST upload the datapoint via WebSocket.
-        this.props.sendMessage(JSON.stringify(datapoint));
+        sendMessage(JSON.stringify(datapoint));
         break;
       default:
         // Remind: Just change the state.
-        this.props.setDatapoint(this.props.deviceId, datapoint);
+        setDatapoint(deviceId, datapoint);
+        break;
     }
   }
   render() {
     const { isMenuShow, target } = this.state;
-    const { device, isLoading, getMessages: t } = this.props;
+    const { deviceId, device, isLoading, getMessages: t } = this.props;
     const { getTarget, onMoreDetailClick, onHide, fetch, eventHandler } = this;
     return (
       <div>
@@ -79,10 +89,10 @@ class DeviceDetail extends React.Component {
                 alignConfig={{ points: ['tr', 'bc'], offset: [20, -20]}}
               >
                 <Menu.Menu key="menu">
-                  <StyledLink to={updatePathname(`/devices/${device.deviceId}/info`)}>
+                  <StyledLink to={updatePathname(`/devices/${deviceId}/info`)}>
                     <Menu.MenuItem>{t('deviceIntro')}</Menu.MenuItem>
                   </StyledLink>
-                  <StyledLink to={updatePathname(`/devices/${device.deviceId}/trigger`)}>
+                  <StyledLink to={updatePathname(`/devices/${deviceId}/trigger`)}>
                     <Menu.MenuItem>{t('triggerAndAction')}</Menu.MenuItem>
                   </StyledLink>
                 </Menu.Menu>
@@ -99,12 +109,12 @@ class DeviceDetail extends React.Component {
               <Container>
                 <CardWrapper>
                   {
-                    (device.datachannels || []).map(c => (
+                    device.datachannels && device.datachannels.map(c => (
                       <DataChannelCard
                         key={c.datachannelId}
                         data-width="half"
                         header={
-                          <StyledLink to={updatePathname(`/devices/${device.deviceId}/dataChannels/${c.datachannelId}`)}>
+                          <StyledLink to={updatePathname(`/devices/${deviceId}/dataChannels/${c.datachannelId}`)}>
                             <CardHeaderIcon>
                               <IconFold />
                             </CardHeaderIcon>
@@ -117,7 +127,7 @@ class DeviceDetail extends React.Component {
                           dataChannelProps={{
                             id: c.datachannelId,
                             type: dataChannelTypeMapper(c.channelType.name, c.type),
-                            values: c.datapoints.values || {},
+                            values: c.datapoints.values,
                             format: c.format,
                           }}
                           eventHandler={eventHandler}