+++ packages/mcs-lite-mobile-web/src/containers/DeviceDetail/index.js
@@ -1,6 +1,7 @@
 /* global window */
 
 import { connect } from 'react-redux';
+import R from 'ramda';
 import compose from 'recompose/compose';
 import { withGetMessages } from 'react-intl-inject-hoc';
 import { connectSocket } from 'mcs-lite-connect';
@@ -10,8 +11,11 @@ import DeviceDetail from './DeviceDetail';
 
 export const mapStateToProps = ({ devices, ui }, { params: { deviceId }}) => ({
   deviceId,
-  device: devices[deviceId],
+  device: R.pathOr({}, [deviceId])(devices),
   isLoading: ui.isLoading,
+
+  // For WebSocket Config
+  deviceKey: R.pathOr(undefined, [deviceId, 'deviceKey'])(devices),
 });
 export const mapDispatchToProps = {
   fetchDeviceDetail: actions.fetchDeviceDetail,
@@ -23,7 +27,8 @@ const wsHost = `ws://${window.location.hostname}:${process.env.REACT_APP_SOCKET_
 export default compose(
   connect(mapStateToProps, mapDispatchToProps),
   connectSocket(
-    ({ device }) => device && `${wsHost}/deviceId/${device.deviceId}/deviceKey/${device.deviceKey}`,
+    ({ deviceId, deviceKey }) => deviceKey
+      && `${wsHost}/deviceId/${deviceId}/deviceKey/${deviceKey}`,
     props => datapoint => props.setDatapoint(props.deviceId, datapoint, true),
     'sendMessage', // propsName
   ),