+++ packages/mcs-lite-ui/src/DataChannelAdapter/DataChannelAdapter.js
@@ -9,7 +9,7 @@ const Wrapper = styled.div`
 
 /**
  * type Event = {
- *   type: 'submit'|'change'|'clear', // event type
+ *   type: 'SUBMIT'|'CHANGE'|'CLEAR', // event type
  *   id: string,                      // data channel id
  *   values: {                        // datapoint values
  *     value: string|number,
@@ -39,7 +39,7 @@ class DataChannelAdapter extends React.Component {
       values: PropTypes.shape({
         value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
         period: PropTypes.number,
-      }).isRequired,
+      }),
       format: PropTypes.shape({
         unit: PropTypes.string,
         items: PropTypes.array,
@@ -53,14 +53,24 @@ class DataChannelAdapter extends React.Component {
   // TODO: refactor these handler codes. (maybe hoist)
   switchByType = () => {
     const { dataChannelProps, eventHandler } = this.props;
-    const { id, values, type, format } = dataChannelProps;
+    const { id, type, format } = dataChannelProps;
+
+    // Remind: API response could be 'null' value.
+    const values = R.defaultTo({})(dataChannelProps.values);
 
     return R.cond([
+
+      /**
+       * Switch
+       *
+       * @author Michael Hsu
+       */
+
       [R.equals('SWITCH_CONTROL'), () =>
         <DataChannel.ControlSwitch
           value={Boolean(values.value)}
           onSubmit={() => eventHandler({
-            type: 'submit',
+            type: 'SUBMIT',
             id,
             values: { value: !values.value ? 1 : 0 },
           })}
@@ -72,84 +82,119 @@ class DataChannelAdapter extends React.Component {
           labels={['OFF', 'ON']}
         />,
       ],
+
+      /**
+       * Integer
+       *
+       * @author Michael Hsu
+       */
+
       [R.equals('INTEGER_CONTROL'), () =>
         <DataChannel.ControlNumber
           placeholder="Integer only"
-          unit={`單位：${format.unit}`}
-          value={values.value ? parseInt(values.value, 10) : ''}
-          onSubmit={() => eventHandler({ type: 'submit', id, values })}
-          onChange={e => eventHandler({ type: 'change', id, values: { value: e.target.value }})}
-          onClear={() => eventHandler({ type: 'clear', id, values: 0 })}
+          unit={format.unit && `單位：${format.unit}`}
+          value={R.isNil(values.value) ? '' : parseInt(values.value, 10)}
+          onSubmit={() => eventHandler({ type: 'SUBMIT', id, values })}
+          onChange={e => eventHandler({ type: 'CHANGE', id, values: { value: e.target.value }})}
+          onClear={() => eventHandler({ type: 'CLEAR', id, values: {}})}
         />,
       ],
       [R.equals('INTEGER_DISPLAY'), () =>
         <DataChannel.DisplayUnitValue
-          value={values.value || NO_DATA_PLACEHOLDER}
+          value={R.isNil(values.value) ? NO_DATA_PLACEHOLDER : parseInt(values.value, 10)}
           unit={format.unit}
         />,
       ],
+
+      /**
+       * Float
+       *
+       * @author Michael Hsu
+       */
+
+      [R.equals('FLOAT_CONTROL'), () =>
+        <DataChannel.ControlNumber
+          placeholder="Float only"
+          unit={format.unit && `單位：${format.unit}`}
+          value={R.isNil(values.value) ? '' : parseFloat(values.value, 10)}
+          onSubmit={() => eventHandler({ type: 'SUBMIT', id, values })}
+          onChange={e => eventHandler({ type: 'CHANGE', id, values: { value: e.target.value }})}
+          onClear={() => eventHandler({ type: 'CLEAR', id, values: {}})}
+        />,
+      ],
+      [R.equals('FLOAT_DISPLAY'), () =>
+        <DataChannel.DisplayUnitValue
+          value={R.isNil(values.value) ? NO_DATA_PLACEHOLDER : parseFloat(values.value, 10)}
+          unit={format.unit}
+        />,
+      ],
+
+      /**
+       * String
+       *
+       * @author Michael Hsu
+       */
+
       [R.equals('STRING_CONTROL'), () =>
         <DataChannel.ControlString
           placeholder="String only"
-          value={values.value || ''}
-          onSubmit={() => eventHandler({ type: 'submit', id, values })}
-          onChange={e => eventHandler({ type: 'change', id, values: { value: e.target.value }})}
-          onClear={() => eventHandler({ type: 'clear', id, values: 0 })}
+          value={R.isNil(values.value) ? '' : values.value}
+          onSubmit={() => eventHandler({ type: 'SUBMIT', id, values })}
+          onChange={e => eventHandler({ type: 'CHANGE', id, values: { value: e.target.value }})}
+          onClear={() => eventHandler({ type: 'CLEAR', id, values: {}})}
         />,
       ],
       [R.equals('STRING_DISPLAY'), () =>
         <DataChannel.DisplayString
           placeholder="String only"
-          value={values.value}
+          value={R.isNil(values.value) ? '' : values.value}
         />,
       ],
+
+      /**
+       * HEX
+       *
+       * @author Michael Hsu
+       */
+
       [R.equals('HEX_CONTROL'), () =>
         <DataChannel.ControlString
           placeholder="Hex only"
-          value={values.value || ''}
-          onSubmit={() => eventHandler({ type: 'submit', id, values })}
-          onChange={e => eventHandler({ type: 'change', id, values: { value: e.target.value }})}
-          onClear={() => eventHandler({ type: 'clear', id, values: 0 })}
+          value={R.isNil(values.value) ? '' : values.value}
+          onSubmit={() => eventHandler({ type: 'SUBMIT', id, values })}
+          onChange={e => eventHandler({ type: 'CHANGE', id, values: { value: e.target.value }})}
+          onClear={() => eventHandler({ type: 'CLEAR', id, values: {}})}
         />,
       ],
       [R.equals('HEX_DISPLAY'), () =>
         <DataChannel.DisplayString
-          placeholder="String only"
-          value={values.value || ''}
+          placeholder="Hex only"
+          value={R.isNil(values.value) ? '' : values.value}
         />,
       ],
-      [R.equals('FLOAT_DISPLAY'), () =>
-        <DataChannel.DisplayUnitValue
-          value={values.value || NO_DATA_PLACEHOLDER}
-          unit={format.unit}
-        />,
-      ],
-      [R.equals('FLOAT_CONTROL'), () =>
-        <DataChannel.ControlNumber
-          placeholder="Float only"
-          unit={`單位：${format.unit}`}
-          value={values.value ? parseFloat(values.value, 10) : ''}
-          onSubmit={() => eventHandler({ type: 'submit', id, values })}
-          onChange={e => eventHandler({ type: 'change', id, values: { value: e.target.value }})}
-          onClear={() => eventHandler({ type: 'clear', id, values: 0 })}
-        />,
-      ],
+
+      /**
+       * GPIO
+       *
+       * @author Michael Hsu
+       */
+
       [R.equals('GPIO_CONTROL'), () => {
         const labels = ['Low', 'High'];
-        const valueMapper = index => index ? labels[index] : NO_DATA_PLACEHOLDER;
+        const valueMapper = index => labels[index] || NO_DATA_PLACEHOLDER;
 
         return (
           <DataChannel.ControlRange
-            value={values.value || ''}
+            value={R.isNil(values.value) ? '' : values.value}
             labels={labels}
             valueMapper={valueMapper}
             onChange={e => eventHandler({
-              type: 'change',
+              type: 'CHANGE',
               id,
               values: { value: e.target.value },
             })}
             onSubmit={() => eventHandler({
-              type: 'submit',
+              type: 'SUBMIT',
               id,
               values: { value: values.value },
             })}
@@ -158,109 +203,136 @@ class DataChannelAdapter extends React.Component {
       }],
       [R.equals('GPIO_DISPLAY'), () =>
         <DataChannel.DisplayStatus
-          value={values.value}
+          value={R.isNil(values.value) ? undefined : values.value}
           labels={['Low', 'High']}
         />,
       ],
+
+      /**
+       * Category
+       *
+       * @author Michael Hsu
+       */
+
       [R.equals('CATEGORY_CONTROL'), () => {
-        const value = values.value
-          ? R.findIndex(R.propEq('value', values.value))(format.items)
-          : undefined;
-        const valueMapper = index => index
+        const value = R.isNil(values.value)
+          ? undefined
+          : R.findIndex(R.propEq('value', values.value))(format.items);
+        const valueMapper = index => format.items[index]
           ? format.items[index].value
           : NO_DATA_PLACEHOLDER;
 
         return (
           <DataChannel.ControlRange
-            value={value || ''}
-            labels={format.items.map(R.prop('name'))}
-            valueMapper={valueMapper}
-            onChange={e => eventHandler({
-              type: 'change',
-              id,
-              values: { value: valueMapper(e.target.value), period: values.period },
-            })}
-            onSubmit={() => eventHandler({
-              type: 'submit',
-              id,
-              values: { value: valueMapper(values.value), period: values.period },
-            })}
-          />
-        );
-      }],
-      [R.equals('CATEGORY_DISPLAY'), () => {
-        const value = values.value
-          ? R.findIndex(R.propEq('value', values.value))(format.items)
-          : undefined;
-
-        return (
-          <DataChannel.DisplayStatus
             value={value}
-            labels={format.items.map(R.prop('name'))}
-          />
-        );
-      }],
-      [R.equals('ANALOG_CONTROL'), () => {
-        const valueMapper = index => index || NO_DATA_PLACEHOLDER;
-
-        return (
-          <DataChannel.ControlRange
-            value={values.value || ''}
+            labels={R.pluck('name')(format.items)}
             valueMapper={valueMapper}
-            labels={[
-              parseFloat(format.lowerbound, 10),
-              parseFloat(format.upperbound, 10),
-            ]}
             onChange={e => eventHandler({
-              type: 'change',
+              type: 'CHANGE',
               id,
-              values: { value: e.target.value },
+              values: { value: valueMapper(e.target.value) },
             })}
             onSubmit={() => eventHandler({
-              type: 'submit',
+              type: 'SUBMIT',
               id,
               values: { value: values.value },
             })}
           />
         );
       }],
+      [R.equals('CATEGORY_DISPLAY'), () => {
+        const value = R.isNil(values.value)
+          ? undefined
+          : R.findIndex(R.propEq('value', values.value))(format.items);
+
+        return (
+          <DataChannel.DisplayStatus
+            value={value}
+            labels={R.pluck('name')(format.items)}
+          />
+        );
+      }],
+
+      /**
+       * Analog (Control only)
+       *
+       * @author Michael Hsu
+       */
+
+      [R.equals('ANALOG_CONTROL'), () => {
+        // Hint: index could be empty string. (For Controlled Component usage.)
+        const valueMapper = index => R.isNil(index) || R.isEmpty(index)
+          ? NO_DATA_PLACEHOLDER
+          : index;
+
+        return (
+          <DataChannel.ControlRange
+            value={R.isNil(values.value) ? '' : values.value}
+            valueMapper={valueMapper}
+            labels={[
+              parseFloat(format.lowerbound, 10),
+              parseFloat(format.upperbound, 10),
+            ]}
+            onChange={e => eventHandler({
+              type: 'CHANGE',
+              id,
+              values: { value: e.target.value },
+            })}
+            onSubmit={() => eventHandler({
+              type: 'SUBMIT',
+              id,
+              values: { value: values.value },
+            })}
+          />
+        );
+      }],
+
+      /**
+       * PWM
+       *
+       * @author Michael Hsu
+       */
+
       [R.equals('PWM_DISPLAY'), () =>
         <DataChannel.DisplayMultipleValue
           items={[
-            { name: 'Value', value: values.value || NO_DATA_PLACEHOLDER },
-            { name: 'Period', value: values.period || NO_DATA_PLACEHOLDER },
+            { name: 'Value', value: R.isNil(values.value) ? NO_DATA_PLACEHOLDER : values.value },
+            { name: 'Period', value: R.isNil(values.period) ? NO_DATA_PLACEHOLDER : values.period },
           ]}
         />,
       ],
       [R.equals('PWM_CONTROL'), () => {
-        const valueMapper = index => index || NO_DATA_PLACEHOLDER;
+        // Hint: index could be empty string. (For Controlled Component usage.)
+        const valueMapper = index => R.isNil(index) || R.isEmpty(index)
+          ? NO_DATA_PLACEHOLDER
+          : index;
 
         return (
           <Wrapper>
             <DataChannel.ControlPeriod
-              value={values.period}
-              onSubmit={() => eventHandler({ type: 'submit', id, values })}
+              value={R.isNil(values.period) ? '' : values.period}
+              onSubmit={() => eventHandler({ type: 'SUBMIT', id, values })}
               onChange={e => eventHandler({
-                type: 'change',
+                type: 'CHANGE',
                 id,
                 values: { period: e.target.value, value: values.value },
               })}
               placeholder="Integer only"
             />
             <DataChannel.ControlRange
-              value={values.value ? parseFloat(values.value, 10) : ''}
+              value={R.isNil(values.value) ? '' : parseFloat(values.value, 10)}
               valueMapper={valueMapper}
               labels={[
                 parseFloat(format.lowerbound, 10),
                 parseFloat(format.upperbound, 10),
               ]}
               onChange={e => eventHandler({
-                type: 'change',
+                type: 'CHANGE',
                 id,
                 values: { value: e.target.value, period: values.period },
               })}
               onSubmit={() => eventHandler({
-                type: 'submit',
+                type: 'SUBMIT',
                 id,
                 values: { value: values.value, period: values.period },
               })}