+++ packages/mcs-lite-ui/src/DataChannelAdapter/__tests__/SWITCH_CONTROL.test.js
@@ -1,0 +1,108 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import { ThemeProvider } from 'styled-components';
+import { theme } from 'mcs-lite-theme';
+import DataChannelAdapter from '../DataChannelAdapter';
+import DataChannel from '../../DataChannel';
+
+it('should render SWITCH_CONTROL correctly with default value to false', () => {
+  const wrapper = mount(
+    <ThemeProvider theme={theme}>
+      <DataChannelAdapter
+        dataChannelProps={{
+          id: 'id',
+          type: 'SWITCH_CONTROL',
+          values: {},
+          format: {},
+        }}
+        eventHandler={() => {}}
+      />
+    </ThemeProvider>,
+  );
+
+  expect(wrapper.find(DataChannelAdapter)).toMatchSnapshot();
+});
+
+it('should render SWITCH_CONTROL correctly with true value', () => {
+  const wrapper = mount(
+    <ThemeProvider theme={theme}>
+      <DataChannelAdapter
+        dataChannelProps={{
+          id: 'id',
+          type: 'SWITCH_CONTROL',
+          values: { value: 1 },
+          format: {},
+        }}
+        eventHandler={() => {}}
+      />
+    </ThemeProvider>,
+  );
+
+  expect(wrapper.find(DataChannelAdapter)).toMatchSnapshot();
+});
+
+it('should render SWITCH_CONTROL correctly with false value', () => {
+  const wrapper = mount(
+    <ThemeProvider theme={theme}>
+      <DataChannelAdapter
+        dataChannelProps={{
+          id: 'id',
+          type: 'SWITCH_CONTROL',
+          values: { value: 0 },
+          format: {},
+        }}
+        eventHandler={() => {}}
+      />
+    </ThemeProvider>,
+  );
+
+  expect(wrapper.find(DataChannelAdapter)).toMatchSnapshot();
+});
+
+it('should handle onSubmit with true value', () => {
+  const mockEventHandler = jest.fn();
+
+  const wrapper = mount(
+    <ThemeProvider theme={theme}>
+      <DataChannelAdapter
+        dataChannelProps={{
+          id: 'id',
+          type: 'SWITCH_CONTROL',
+          values: { value: 0 },
+          format: {},
+        }}
+        eventHandler={mockEventHandler}
+      />
+    </ThemeProvider>,
+  );
+
+  expect(mockEventHandler).not.toHaveBeenCalled();
+  wrapper.find(DataChannel.ControlSwitch).props().onSubmit();
+  expect(mockEventHandler).toHaveBeenCalledWith({
+    id: 'id', type: 'SUBMIT', values: { value: 1 }},
+  );
+});
+
+it('should handle onSubmit with false value', () => {
+  const mockEventHandler = jest.fn();
+
+  const wrapper = mount(
+    <ThemeProvider theme={theme}>
+      <DataChannelAdapter
+        dataChannelProps={{
+          id: 'id',
+          type: 'SWITCH_CONTROL',
+          values: { value: 1 },
+          format: {},
+        }}
+        eventHandler={mockEventHandler}
+      />
+    </ThemeProvider>,
+  );
+
+  expect(mockEventHandler).not.toHaveBeenCalled();
+  wrapper.find(DataChannel.ControlSwitch).props().onSubmit();
+  expect(mockEventHandler).toHaveBeenCalledWith({
+    id: 'id', type: 'SUBMIT', values: { value: 0 }},
+  );
+});